ATR-507: refactoring - created separate helper type and subnamespace and class for syntax trivia extension methods
